+++ packages/plugin-hardhat/src/utils/deploy-impl.ts
@@ -99,18 +99,28 @@ async function deployImpl(
 ): Promise<DeployedImpl> {
   const layout = deployData.layout;
 
+  if (opts.useDeployedImplementation && opts.redeployImplementation !== undefined) {
+    throw new UpgradesError(
+      'The useDeployedImplementation and redeployImplementation options cannot both be set at the same time',
+    );
+  }
+
+  const merge = deployData.fullOpts.redeployImplementation === 'always';
+
   const deployment = await fetchOrDeployGetDeployment(
     deployData.version,
     deployData.provider,
     async () => {
       const abi = ImplFactory.interface.format(FormatTypes.minimal) as string[];
       const attemptDeploy = () => {
-        if (opts.useDeployedImplementation) {
-          throw new UpgradesError(
-            'The implementation contract was not previously deployed.',
-            () =>
-              'The useDeployedImplementation option was set to true but the implementation contract was not previously deployed on this network.',
-          );
+        if (deployData.fullOpts.useDeployedImplementation || deployData.fullOpts.redeployImplementation === 'never') {
+          throw new UpgradesError('The implementation contract was not previously deployed.', () => {
+            if (deployData.fullOpts.useDeployedImplementation) {
+              return 'The useDeployedImplementation option was set to true but the implementation contract was not previously deployed on this network.';
+            } else {
+              return "The redeployImplementation option was set to 'never' but the implementation contract was not previously deployed on this network.";
+            }
+          });
         } else {
           return deploy(hre, opts, ImplFactory, ...deployData.fullOpts.constructorArgs);
         }
@@ -119,7 +129,7 @@ async function deployImpl(
       return { ...deployment, layout };
     },
     opts,
-    undefined,
+    merge,
     remoteDeploymentId => getRemoteDeployment(hre, remoteDeploymentId),
   );