+++ packages/plugin-truffle/src/utils/deploy-impl.ts
@@ -88,22 +88,38 @@ function encodeArgs(Contract: ContractClass, constructorArgs: unknown[]): string
 async function deployImpl(deployData: DeployData, Contract: ContractClass, opts: UpgradeOptions): Promise<any> {
   const layout = deployData.layout;
 
-  const impl = await fetchOrDeploy(deployData.version, deployData.provider, async () => {
-    const abi = (Contract as any).abi;
-    const attemptDeploy = () => {
-      if (opts.useDeployedImplementation) {
-        throw new UpgradesError(
-          'The implementation contract was not previously deployed.',
-          () =>
-            'The useDeployedImplementation option was set to true but the implementation contract was not previously deployed on this network.',
-        );
-      } else {
-        return deploy(deployData.fullOpts.deployer, Contract, ...deployData.fullOpts.constructorArgs);
-      }
-    };
-    const deployment = Object.assign({ abi }, await attemptDeploy());
-    return { ...deployment, layout };
-  });
+  if (opts.useDeployedImplementation && opts.redeployImplementation !== undefined) {
+    throw new UpgradesError(
+      'The useDeployedImplementation and redeployImplementation options cannot both be set at the same time',
+    );
+  }
+
+  const merge = deployData.fullOpts.redeployImplementation === 'always';
+
+  const impl = await fetchOrDeploy(
+    deployData.version,
+    deployData.provider,
+    async () => {
+      const abi = (Contract as any).abi;
+      const attemptDeploy = () => {
+        if (deployData.fullOpts.useDeployedImplementation || deployData.fullOpts.redeployImplementation === 'never') {
+          throw new UpgradesError('The implementation contract was not previously deployed.', () => {
+            if (deployData.fullOpts.useDeployedImplementation) {
+              return 'The useDeployedImplementation option was set to true but the implementation contract was not previously deployed on this network.';
+            } else {
+              return "The redeployImplementation option was set to 'never' but the implementation contract was not previously deployed on this network.";
+            }
+          });
+        } else {
+          return deploy(deployData.fullOpts.deployer, Contract, ...deployData.fullOpts.constructorArgs);
+        }
+      };
+      const deployment = Object.assign({ abi }, await attemptDeploy());
+      return { ...deployment, layout };
+    },
+    opts,
+    merge,
+  );
 
   return { impl, kind: opts.kind };
 }